wgpu - removed redundant GpuTextureView constructor

+++ src/Friflo.WGPU/API/RenderFrame.cs
@@ -39,7 +39,7 @@ public static partial class WgpuExtensions
                 return new RenderTarget(default, null, surfaceTexture.status, null, targetSize);
             }
             var handle = wgpuTextureCreateView(surfaceTexture.texture, null);
-            var view = new GpuTextureView(handle, null);
+            var view = new GpuTextureView((nint)handle, null);
         
             fixed (byte* labelPtr = encoderLabel) {
                 var label = WgpuUtils.FromPtrSpan(labelPtr, encoderLabel);
